chore(folder-structure): Move i18n to libs
