+++ src/app/profiles/index.tsx
@@ -1,7 +1,7 @@
 import './browser';
 import React from 'react';
 import { render } from 'react-dom';
-import i18n, { options } from 'i18n';
+import i18n, { options } from 'libs/i18n';
 import App from './App';
 
 i18n
