+++ src/libs/i18n/resources/fr/extension.json
