+++ src/libs/i18n/resources/fr/index.ts
