+++ src/libs/i18n/resources/fr/profiles.json
