+++ src/libs/i18n/resources/index.ts
